creature: move Creature_AIInfo to trx
