+++ src/libtrx/game/creature/common.c
@@ -1,5 +1,43 @@
 #include "game/creature.h"
+#include "game/lara/common.h"
 #include "game/random.h"
+#include "game/rooms.h"
+#include "utils.h"
+
+#define MAX_CREATURE_DISTANCE (WALL_L * 30)
+
+static ITEM *M_GetEnemy(const ITEM *item);
+
+#if TR_VERSION == 2
+extern void Creature_GetBaddieTarget(int16_t item_num, bool goody);
+#endif
+
+static ITEM *M_GetEnemy(const ITEM *const item)
+{
+    CREATURE *const creature = item->data;
+    switch (item->object_id) {
+#if TR_VERSION == 2
+    case O_BANDIT_1:
+    case O_BANDIT_2:
+        Creature_GetBaddieTarget(creature->item_num, false);
+        break;
+
+    case O_MONK_1:
+    case O_MONK_2:
+        Creature_GetBaddieTarget(creature->item_num, true);
+        break;
+#endif
+
+    default:
+        creature->enemy = Lara_GetItem();
+        break;
+    }
+
+    if (creature->enemy != nullptr) {
+        return creature->enemy;
+    }
+    return Lara_GetItem();
+}
 
 void Creature_Initialise(const int16_t item_num)
 {
@@ -22,4 +60,60 @@ bool Creature_Activate(const int16_t item_num)
 
     item->status = IS_ACTIVE;
     return true;
+}
+
+void Creature_AIInfo(ITEM *const item, AI_INFO *const info)
+{
+    CREATURE *const creature = item->data;
+    if (creature == nullptr) {
+        return;
+    }
+
+    ITEM *const enemy = M_GetEnemy(item);
+    const int16_t *const zone = Box_GetLotZone(&creature->lot);
+
+    {
+        const ROOM *const room = Room_Get(item->room_num);
+        item->box_num =
+            Room_GetWorldSector(room, item->pos.x, item->pos.z)->box;
+        info->zone_num = zone[item->box_num];
+    }
+
+    {
+        const ROOM *const room = Room_Get(enemy->room_num);
+        enemy->box_num =
+            Room_GetWorldSector(room, enemy->pos.x, enemy->pos.z)->box;
+        info->enemy_zone_num = zone[enemy->box_num];
+    }
+
+    if (((Box_GetBox(enemy->box_num)->overlap_index & creature->lot.block_mask)
+         != 0)
+        || (creature->lot.node[item->box_num].search_num
+            == (creature->lot.search_num | BOX_BLOCKED_SEARCH))) {
+        info->enemy_zone_num |= BOX_BLOCKED;
+    }
+
+    const OBJECT *const obj = Object_Get(item->object_id);
+    const int32_t z = enemy->pos.z
+        - ((obj->pivot_length * Math_Cos(item->rot.y)) >> W2V_SHIFT)
+        - item->pos.z;
+    const int32_t x = enemy->pos.x
+        - ((obj->pivot_length * Math_Sin(item->rot.y)) >> W2V_SHIFT)
+        - item->pos.x;
+    const int16_t angle = Math_Atan(z, x);
+
+    if (creature->enemy == nullptr) {
+        info->distance = 0x7FFFFFFF;
+    } else if (
+        ABS(x) > MAX_CREATURE_DISTANCE || ABS(z) > MAX_CREATURE_DISTANCE) {
+        info->distance = SQUARE(MAX_CREATURE_DISTANCE);
+    } else {
+        info->distance = SQUARE(x) + SQUARE(z);
+    }
+
+    info->angle = angle - item->rot.y;
+    info->enemy_facing = angle - enemy->rot.y + DEG_180;
+    info->ahead = info->angle > -FRONT_ARC && info->angle < FRONT_ARC;
+    info->bite = info->ahead && ABS(enemy->pos.y - item->pos.y) <= STEP_L
+        && (TR_VERSION == 1 || enemy->hit_points > 0);
 }
