+++ src/tr2/game/creature.h
@@ -13,7 +13,7 @@ int32_t Creature_Vault(
     int16_t item_num, int16_t angle, int32_t vault, int32_t shift);
 void Creature_Kill(
     ITEM *item, int32_t kill_anim, int32_t kill_state, int32_t lara_kill_state);
-void Creature_GetBaddieTarget(int16_t item_num, int32_t goody);
+void Creature_GetBaddieTarget(int16_t item_num, bool goody);
 int32_t Creature_CanTargetEnemy(const ITEM *item, const AI_INFO *info);
 bool Creature_IsAlly(const ITEM *item);
 int32_t Creature_ShootAtLara(
